Chain dependency downloading to force reliable execution order
Dojo's require downloads all dependencies concurrently, and executes each as
part of finalizing it. This can cause the test specs to be processed in a
different order. If using a test framework that supports test ordering (like
jasmine), it needs the specs to come in the same order. Otherwise, the ordering
will produce different ordering, even for the same random seed.